+++ src/views/Settings.vue
@@ -10,16 +10,27 @@
         <UserSettings></UserSettings>
       </v-col>
     </v-row>
+    <v-row v-if="isStaff">
+      <v-col sm="12">
+        <RetentionCard></RetentionCard>
+      </v-col>
+    </v-row>
   </v-container>
 </template>
 
 <script>
   import CommonSettings from "../components/CommonSettings"
   import UserSettings from "../components/UserSettings"
+  import RetentionCard from "../components/RetentionCard"
 
   export default {
     name: "Settings",
-    components: { UserSettings, CommonSettings },
+    components: { UserSettings, CommonSettings, RetentionCard },
+    computed: {
+      isStaff() {
+        return this.$store.getters.isStaff
+      },
+    },
   }
 </script>